+++ server/session.go
@@ -5,7 +5,6 @@ import (
 	"github.com/pkg/sftp"
 	"slider/pkg/instance"
 	"slider/pkg/interpreter"
-	"slider/pkg/sio"
 	"slider/pkg/slog"
 	"strings"
 	"sync"
@@ -257,54 +256,6 @@ func (session *Session) replyConnRequest(request *ssh.Request, ok bool, payload 
 		pMsg,
 	)
 	return request.Reply(ok, payload)
-}
-
-func (session *Session) uploadFile(src, dst string) <-chan sio.Status {
-	status := make(chan sio.Status)
-	fileList, action := sio.NewFileAction(session.sshConn, src, dst)
-	go func() {
-		for s := range action.UploadToClient(fileList) {
-			status <- s
-		}
-		close(status)
-	}()
-
-	return status
-}
-
-func (session *Session) downloadFile(src, dst string) <-chan sio.Status {
-	status := make(chan sio.Status)
-	fileList, action := sio.NewFileAction(session.sshConn, src, dst)
-	go func() {
-		for s := range action.DownloadFromClient(fileList) {
-			status <- s
-		}
-		close(status)
-	}()
-
-	return status
-}
-
-func (session *Session) downloadFileBatch(fileListPath string) <-chan sio.Status {
-	status := make(chan sio.Status)
-	fileList, action, err := sio.NewBatchAction(session.sshConn, fileListPath)
-	if err != nil {
-		status <- sio.Status{
-			FileInfo: sio.FileInfo{},
-			Success:  false,
-			Err:      fmt.Errorf("failed to create batch action"),
-		}
-		close(status)
-		return status
-	}
-	go func() {
-		for s := range action.DownloadFromClient(fileList) {
-			status <- s
-		}
-		close(status)
-	}()
-
-	return status
 }
 
 func (session *Session) socksEnable(port int, exposePort bool) {
